refactor: extract selectWord event handler
단어를 선택하는 이벤트 파일로 추출
